Create rechtspraak_api
This is a short python script to connect with the RESTful API which gives access to the official website of the Dutch Judiciary (www.rechtspraak.nl) and execute a search query. Using search parameters, you can determine the period, the area of law and the relevant institution(s). Feel free to comment and improve!